fix: fix errors on newer ASUS laptop models
On newer ASUS laptop models, the charging threshold variable can be
exposed under different names i.e. BAT0, BAT1, BATC, and BATT. This
would cause an error in the previous version of the application which
used the single hard-coded value BAT0.

Revert to using `/bin/bash` for writing to the systemd service that
persists the charging threshold level as `/bin/sh` does not seem to
function as intended.

Reference
- https://github.com/torvalds/linux/blob/1797d588af15174d4a4e7159dac8c800538e4f8c/drivers/platform/x86/asus-wmi.c#L440-L448

Fixes
- https://github.com/leveson/bat/issues/1