refactor: update function that queries Bash path
Change the function that queries the location of the Bash shell on the
operating system to look inside of the $PATH environment variable
instead of defaulting to standard locations such as /bin/ and /usr/bin/
that maybe not be true for all Unix operating systems i.e. NixOS.

See https://github.com/tshakalekholoane/bat/issues/26.