fix: restore persistence functionality
This commit addresses an issue where the switch to using systemd
templates and format specifiers introduced a regression, causing the
persistence functionality to no longer work as expected.

To resolve this issue, we have reverted to an earlier implementation
that manually writes all the units and uses Go's exec.LookPath to
resolve the shell path.

Resolves: #65